Basic migration of Sentiment Analysis to v0.7. Still not using "common code"
